[refactor] Adapt `common/rpc` for new structure
